+++ src/main/java/edu/indiana/sice/dscspidal/mpicommonio/SparseVector.java
@@ -33,8 +33,8 @@ public class SparseVector implements SmallVector<Double>, Sparse {
     }
 
     @Override
-    public int size() {
-        return n;
+    public long size() {
+        return (long) n;
     }
 
     @Override
